pyproject.toml
[project]
name = "jarvis-recipes"
version = "0.1.0"
description = "Jarvis Recipes FastAPI service"
authors = [
    {name = "alexberardi",email = "alexmberardi@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.123.9,<0.124.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "lxml (>=5.3.0,<6.0.0)",
    "pytesseract (>=0.3.10,<0.4.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "Pillow (>=10.4.0,<11.0.0)",
    "pillow-heif (>=0.16.0,<0.17.0)",
    "boto3 (>=1.34.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.3.0,<9.0.0)",
    "pytest-asyncio (==0.23.8)"
]

[tool.poetry]
packages = [{ include = "jarvis_recipes" }]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "asyncio: mark test as asyncio-based",
]
